repository: Product

+++ E-Commerce/src/main/java/com/springproject/ecommerce/repository/ProductRepository.java
@@ -8,5 +8,5 @@ import com.springproject.ecommerce.model.Product;
 
 public interface ProductRepository extends JpaRepository<Product, Integer>{
 
-	List<Product> findProductByCategoryId(Integer categoryId);
+	List<Product> findProductByCategoryId(Long categoryId);
 }
